Refactored json import getUploadFileFns
