+++ client/src/app/site/pages/meetings/pages/motions/pages/workflows/components/workflow-import/workflow-import.component.ts
@@ -1,13 +1,11 @@
 import { Location } from '@angular/common';
 import { Component } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { marker as _ } from '@biesbjerg/ngx-translate-extract-marker';
 import { TranslateService } from '@ngx-translate/core';
+import { getUploadFileJson } from 'src/app/infrastructure/utils/import/json-import-file-utils';
 import { FileData } from 'src/app/ui/modules/file-upload/components/file-upload/file-upload.component';
 
 import { MotionWorkflowControllerService } from '../../../../modules/workflows/services';
-
-export const WRONG_JSON_IMPORT_FORMAT_ERROR_MSG = _(`Import data needs to have the JSON format`);
 
 @Component({
     selector: `os-workflow-import`,
@@ -28,25 +26,7 @@ export class WorkflowImportComponent {
 
     public getUploadFileFn(): (file: FileData) => any {
         return async file => {
-            const workflowJson = await new Promise<any[]>(resolve => {
-                const reader = new FileReader();
-                reader.addEventListener(`load`, progress => {
-                    let result;
-                    try {
-                        result = JSON.parse(progress.target!.result as string);
-                    } catch (e) {
-                        this.snackbar.open(
-                            `${this.translate.instant(`Error`)}: ${this.translate.instant(
-                                WRONG_JSON_IMPORT_FORMAT_ERROR_MSG
-                            )}`,
-                            `OK`
-                        );
-                        throw new Error(WRONG_JSON_IMPORT_FORMAT_ERROR_MSG);
-                    }
-                    resolve(result);
-                });
-                reader.readAsText(file.mediafile);
-            });
+            const workflowJson = await getUploadFileJson<any[]>(file, this.snackbar);
             return this.repo.import(workflowJson);
         };
     }
