+++ client/src/app/site/pages/organization/pages/committees/pages/committee-detail/modules/committee-detail-meeting/components/meeting-import/meeting-import.component.ts
@@ -5,8 +5,8 @@ import { TranslateService } from '@ngx-translate/core';
 import { Id } from 'src/app/domain/definitions/key-types';
 import { Identifiable } from 'src/app/domain/interfaces';
 import { ImportMeeting } from 'src/app/gateways/repositories/meeting-repository.service';
+import { getUploadFileJson } from 'src/app/infrastructure/utils/import/json-import-file-utils';
 import { BaseComponent } from 'src/app/site/base/base.component';
-import { WRONG_JSON_IMPORT_FORMAT_ERROR_MSG } from 'src/app/site/pages/meetings/pages/motions/pages/workflows/components/workflow-import/workflow-import.component';
 import { MeetingControllerService } from 'src/app/site/pages/meetings/services/meeting-controller.service';
 import { ComponentServiceCollectorService } from 'src/app/site/services/component-service-collector.service';
 import { OpenSlidesRouterService } from 'src/app/site/services/openslides-router.service';
@@ -47,25 +47,7 @@ export class MeetingImportComponent extends BaseComponent implements OnInit {
 
     public getUploadFileFn(): (file: FileData) => Promise<Identifiable> {
         return async file => {
-            const meeting = await new Promise<ImportMeeting>(resolve => {
-                const reader = new FileReader();
-                reader.addEventListener(`load`, progress => {
-                    let result;
-                    try {
-                        result = JSON.parse(progress.target.result as string);
-                    } catch (e) {
-                        this.snackbar.open(
-                            `${this.translate.instant(`Error`)}: ${this.translate.instant(
-                                WRONG_JSON_IMPORT_FORMAT_ERROR_MSG
-                            )}`,
-                            `OK`
-                        );
-                        throw new Error(WRONG_JSON_IMPORT_FORMAT_ERROR_MSG);
-                    }
-                    resolve(result);
-                });
-                reader.readAsText(file.mediafile);
-            });
+            const meeting = await getUploadFileJson<ImportMeeting>(file, this.snackbar);
             return this.repo.import(this._committeeId, meeting);
         };
     }
